requirements.txt
flask==3.0.2
python-chess==1.999